Rename the component structure to match Semantic naming
Rather than introduce a second naming/organization scheme, update all the component directories to work as closely 1:1 with Semantic UI components as possible.
