+++ src/shared/components/collections/index.ts
@@ -1,4 +1,4 @@
 /**
- * Register Organisms here
+ * Register collections here
  */
 // export { default as Example } from "./Example/Example";
