+++ src/shared/components/index.ts
@@ -1,8 +1,9 @@
 /**
  * Register components here
  */
-export * from "./atoms";
-export * from "./molecules";
-// export * from "./organisms";
+export * from "./elements";
+export * from "./modules";
+// export * from "./collections";
+// export * from "./views";
 export * from "./pages";
 export * from "./templates";
