+++ src/shared/components/modules/MainMenu/MainMenu.stories.tsx
