+++ src/shared/components/modules/MainMenu/MainMenu.test.tsx
