+++ src/shared/components/modules/MainMenu/MainMenu.tsx
