+++ src/shared/components/modules/MainMenu/README.md
