+++ src/shared/components/modules/MainMenu/index.ts
