+++ src/shared/components/modules/index.ts
@@ -1,5 +1,5 @@
 /**
- * Register Molecules here
+ * Register modules here
  */
 export { MainMenu } from "./MainMenu";
 export { MenuItemArray, MenuItem } from "./MenuItemArray";
