+++ src/shared/components/views/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Register views here
+ */
+// export { default as Example } from "./Example/Example";
